Add Fortran interoperability layer
This code wraps a Fortran BMI and allows it to be compiled and linked
as a C library. The functions in this layer are what will be called
from Cython and Python.

I've chosen to `use` all libraries listed in the **plugin.yaml** file.

The `plugin_module` variable matches the derived type that holds the
BMI for a Fortran model (e.g., `bmi_heat`). I've chosen to use only
the first of the possible entry points to define this variable. I
think this is OK, since if more than one entry point is given, then
that's a bigger problem, since I've only set up the interoperability
layer to handle one BMI-ed model at a time.