Update variable in loop
It was fascinating to learn that globally scoped variables in jinja2
can't be updated inside a loop. Here, the max length of a component
name wasn't being saved. I found a solution that neatly fit the
problem I was seeing: https://stackoverflow.com/a/42888347/1563298.
Thank you stackoverflow user Paddy V!